pyproject.toml
[project]
name = "cat-dog-recog"
version = "0.1.0"
description = "Cat/Dog Image Classification with PyTorch"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "requests>=2.32.5",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "scikit-learn>=1.6.0",
    "pillow>=10.4.0",
    "seaborn>=0.13.2",
    "ruff>=0.14.10",
    "torchinfo>=1.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
